+++ public/webgpu/visibility/scenes/visibility.js
@@ -13,7 +13,7 @@ const polytopeGeometry = new PolytopeGeometry({
 		2, -1, 0,
 		3, 0, 0,
 	),
-	indices: Uint8Array.of(
+	indices: Uint32Array.of(
 		0, 1, 2,
 		1, 3, 2,
 		1, 4, 3,
